Unitstacks Fix
I fixed some of the obvious issues for now, but I think for a long-term solution, we might wanna do a full snap. Click that "Auto in all" button and kiss all the VP locations goodbye.

Because these don't only exist for the heightmap changes, there's also a lot of border and river ones I've missed, and it's not possible to make all of the odd unitstack locations go away, without getting forced into pushing that shiny, evil button, by out own greed of time-efficiency.